# docs/design/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BusySlot": {
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "label": {
            "default": "予約済み",
            "title": "Label",
            "type": "string"
          },
          "reservation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "BusySlot",
        "type": "object"
      },
      "CancelInput": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "CancelInput",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "Reservation": {
        "additionalProperties": false,
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "purpose": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          },
          "resource_id": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Resource Id",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "status": {
            "enum": [
              "confirmed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "resource_id",
          "start_at",
          "end_at",
          "purpose",
          "id",
          "subject",
          "status",
          "version"
        ],
        "title": "Reservation",
        "type": "object"
      },
      "Resource": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "room",
            "enum": [
              "room",
              "equipment"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id",
          "active",
          "version"
        ],
        "title": "Resource",
        "type": "object"
      },
...
      "ResourceInput": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "default": "room",
            "enum": [
              "room",
              "equipment"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ResourceInput",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SlotKeeper",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reservations/{reservation_id}/cancel": {
      "post": {
        "description": "所有者、公開版、状態と開始時刻を検査して取消と履歴を確定する。",
        "operationId": "reservations_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Endpoint",
        "tags": [
          "reservations"
        ]
      }
    },
    "/resources": {
      "get": {
        "description": "認証済み利用者へ固定順とページ単位で資源を返す。",
        "operationId": "resources_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  },
                  "title": "Response Resources List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Endpoint",
        "tags": [
          "resources"
        ]
      },
      "post": {
        "description": "管理者権限を確認して新しい資源を登録する。",
        "operationId": "resources_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Endpoint",
        "tags": [
          "resources"
        ]
      }
    },
...
    "/resources/{resource_id}/schedule": {
      "get": {
        "description": "一般利用者には他人の識別情報と目的を返さない。",
        "operationId": "schedule",
        "parameters": [
          {
            "in": "path",
            "name": "resource_id",
            "required": true,
            "schema": {
              "title": "Resource Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BusySlot"
                  },
                  "title": "Response Schedule",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Endpoint",
        "tags": [
          "resources"
        ]
      }
    }
  }
}